Encryption, main, and updated compression python scripts.
The scripts all work with .csv files as both input and output, which made the different sub-subsystems easier to design. They work on the IMU data provided on Vula and the decompressed file at the output matches that of the given input data file. The only remaining changes necessary to the code is that used for testing purposes and directory/file changes when converting over to the Raspberry Pi.